refactor: Move some standalone functions to `utils.js`
 - declare as functions first, so other utils methods can use them too
